% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{Constructor Function to Perform Student's T-Test}
\usage{
myttest(x = numeric(), y = numeric(), alpha = double(), paired = boolean())
}
\arguments{
\item{x}{samples of population A}

\item{y}{samples of population B}

\item{alpha}{level of significance}

\item{paired}{whether x and y are paired samples}
}
\value{
list of confidence interval, test type, p-value, decision to reject or not to reject null hypothesis
}
\description{
Constructor Function to Perform Student's T-Test
}
\examples{
set.seed(32); x=rnorm(30,mean=10,sd=15)
set.seed(35); y=rnorm(30,mean=8,sd=15)
ans1 = myttest(x, y, alpha = 0.05, paired=FALSE)
print(ans1)
plot(ans1)

}
